Hi Thomas. When I look at this file in markdrop.com, the file looks great, however, when I view it in smartgit, it looks less than stellar.
My links were not the proper colours and did not necessarily automatically underline.

There was a lot of markdown missing from the suggested link, and the code that I found on other sites that viewed well in markdrop do not look good here.

I am quite pleases with my small image however, that took a bit of doing!